+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.11.28',
+  '2023.12.06',
   // Fallback runtime versions.
+  '2023.11.28',
   '2023.11.21',
-  '2023.11.16',
 ];
 
 /// Sets the current runtime versions.
@@ -62,10 +62,10 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.2.0';
-final String toolStableDartSdkVersion = '3.2.1';
-final String toolStableFlutterSdkVersion = '3.16.1';
-final String toolPreviewDartSdkVersion = '3.3.0-166.0.dev';
-final String toolPreviewFlutterSdkVersion = '3.17.0-0.0.pre';
+final String toolStableDartSdkVersion = '3.2.2';
+final String toolStableFlutterSdkVersion = '3.16.2';
+final String toolPreviewDartSdkVersion = '3.3.0-174.2.beta';
+final String toolPreviewFlutterSdkVersion = '3.18.0-0.1.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);